ai-remix-service/ai_processor.py
import asyncio
from PIL import Image, ImageEnhance, ImageFilter
import io
import numpy as np
from typing import Dict, Any
import torch
from diffusers import StableDiffusionImg2ImgPipeline, StableDiffusionPipeline
from transformers import pipeline
import cv2


class StyleTransferProcessor:
    """Handles style transfer using diffusion models"""

    def __init__(self):
        self.pipeline = None
        self.device = "cuda" if torch.cuda.is_available() else "cpu"

    async def process(
        self, image_data: bytes, style_preset: str, parameters: Dict[str, Any]
    ) -> Image.Image:
        """Apply style transfer to image"""
        # Load image
        image = Image.open(io.BytesIO(image_data))

        # For now, implement a simple filter-based style transfer
        # In production, you'd use a proper diffusion model
        if style_preset == "vintage":
            return self._apply_vintage_style(image)
        elif style_preset == "cartoon":
            return self._apply_cartoon_style(image)
        elif style_preset == "sketch":
            return self._apply_sketch_style(image)
        else:
            return self._apply_generic_style(image, style_preset)

    def _apply_vintage_style(self, image: Image.Image) -> Image.Image:
        """Apply vintage style using filters"""
        # Convert to RGB if needed
        if image.mode != "RGB":
            image = image.convert("RGB")

        # Apply vintage effects
        enhancer = ImageEnhance.Color(image)
        image = enhancer.enhance(0.8)  # Reduce saturation

        enhancer = ImageEnhance.Contrast(image)
        image = enhancer.enhance(1.2)  # Increase contrast

        enhancer = ImageEnhance.Brightness(image)
        image = enhancer.enhance(0.9)  # Slightly darker

        # Add slight blur
        image = image.filter(ImageFilter.GaussianBlur(radius=0.5))

        return image

    def _apply_cartoon_style(self, image: Image.Image) -> Image.Image:
        """Apply cartoon style"""
        # Convert to numpy array
        img_array = np.array(image)

        # Convert to grayscale for edge detection
        gray = cv2.cvtColor(img_array, cv2.COLOR_RGB2GRAY)

        # Apply bilateral filter to reduce noise while preserving edges
        bilateral = cv2.bilateralFilter(img_array, 9, 75, 75)

        # Detect edges
        edges = cv2.adaptiveThreshold(
            gray, 255, cv2.ADAPTIVE_THRESH_MEAN_C, cv2.THRESH_BINARY, 9, 2
        )

        # Combine bilateral filter with edges
        cartoon = cv2.bitwise_and(bilateral, bilateral, mask=edges)

        return Image.fromarray(cartoon)

    def _apply_sketch_style(self, image: Image.Image) -> Image.Image:
        """Apply sketch style"""
        # Convert to numpy array
        img_array = np.array(image)

        # Convert to grayscale
        gray = cv2.cvtColor(img_array, cv2.COLOR_RGB2GRAY)

        # Invert the image
        inverted = 255 - gray

        # Apply Gaussian blur
        blurred = cv2.GaussianBlur(inverted, (21, 21), 0)

        # Invert the blurred image
        inverted_blurred = 255 - blurred

        # Create sketch effect
        sketch = cv2.divide(gray, inverted_blurred, scale=256.0)

        return Image.fromarray(sketch)

    def _apply_generic_style(self, image: Image.Image, style_name: str) -> Image.Image:
        """Apply a generic style based on name"""
        # Default to vintage style
        return self._apply_vintage_style(image)


class BackgroundRemovalProcessor:
    """Handles background removal"""

    def __init__(self):
        self.device = "cuda" if torch.cuda.is_available() else "cpu"

    async def process(
        self, image_data: bytes, parameters: Dict[str, Any]
    ) -> Image.Image:
        """Remove background from image"""
        # Load image
        image = Image.open(io.BytesIO(image_data))

        # For now, implement a simple chroma key removal
        # In production, you'd use a proper segmentation model
        return self._remove_background_simple(image)

    def _remove_background_simple(self, image: Image.Image) -> Image.Image:
        """Simple background removal using color thresholding"""
        # Convert to RGBA
        if image.mode != "RGBA":
            image = image.convert("RGBA")

        # Get image data
        data = np.array(image)

        # Create a mask for white/light backgrounds
        # This is a simple approach - in production use proper segmentation
        r, g, b, a = data[:, :, 0], data[:, :, 1], data[:, :, 2], data[:, :, 3]

        # Create mask for light backgrounds
        mask = (r > 200) & (g > 200) & (b > 200)

        # Set alpha channel based on mask
        data[:, :, 3] = np.where(mask, 0, 255)

        return Image.fromarray(data)


class ImageEnhancer:
    """Handles image enhancement"""

    def __init__(self):
        pass

    async def process(
        self, image_data: bytes, enhancement_type: str, parameters: Dict[str, Any]
    ) -> Image.Image:
        """Enhance image"""
        # Load image
        image = Image.open(io.BytesIO(image_data))

        if enhancement_type == "sharpness":
            return self._enhance_sharpness(image, parameters)
        elif enhancement_type == "color":
            return self._enhance_color(image, parameters)
        elif enhancement_type == "brightness":
            return self._enhance_brightness(image, parameters)
        else:
            return self._enhance_general(image, parameters)

    def _enhance_sharpness(
        self, image: Image.Image, parameters: Dict[str, Any]
    ) -> Image.Image:
        """Enhance image sharpness"""
        factor = parameters.get("strength", 1.5)
        enhancer = ImageEnhance.Sharpness(image)
        return enhancer.enhance(factor)

    def _enhance_color(
        self, image: Image.Image, parameters: Dict[str, Any]
    ) -> Image.Image:
        """Enhance image color"""
        factor = parameters.get("strength", 1.2)
        enhancer = ImageEnhance.Color(image)
        return enhancer.enhance(factor)

    def _enhance_brightness(
        self, image: Image.Image, parameters: Dict[str, Any]
    ) -> Image.Image:
        """Enhance image brightness"""
        factor = parameters.get("strength", 1.1)
        enhancer = ImageEnhance.Brightness(image)
        return enhancer.enhance(factor)

    def _enhance_general(
        self, image: Image.Image, parameters: Dict[str, Any]
    ) -> Image.Image:
        """Apply general enhancement"""
        # Apply multiple enhancements
        factor = parameters.get("strength", 1.2)

        enhancer = ImageEnhance.Contrast(image)
        image = enhancer.enhance(factor)

        enhancer = ImageEnhance.Color(image)
        image = enhancer.enhance(factor)

        enhancer = ImageEnhance.Sharpness(image)
        image = enhancer.enhance(factor)

        return image
