ai-remix-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
Pillow==10.1.0
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
diffusers==0.24.0
transformers==4.35.0
accelerate==0.24.1
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0 